backend/appointment-service/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pyjwt
bcrypt
pydantic
